Former-commit-id: 264400dcaa906651fdd20fe4bf1d011d6c6c3da4

+++ cocos2d/CCParticleSystemQuad.m
@@ -190,7 +190,7 @@
 		CC_SWAP( top, bottom);
 		
 		ccV3F_C4B_T2F_Quad *quadCollection; 
-		uint start, end; 
+		NSUInteger start, end; 
 		if (useBatchNode_)
 		{
 			quadCollection = [[batchNode_ textureAtlas] quads]; 
@@ -204,7 +204,7 @@
 			end = totalParticles; 
 		}
 
-		for(int i=start; i<end; i++) {
+		for(NSInteger i=start; i<end; i++) {
 			// bottom-left vertex:
 			quadCollection[i].bl.texCoords.u = left;
 			quadCollection[i].bl.texCoords.v = bottom;
